# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Garage API",
        "description": "Documentation de l'API Gestion Garage",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/reparations": {
            "get": {
                "tags": [
                    "Reparations"
                ],
                "summary": "Lister les reparations",
                "operationId": "4f69f77028e02a49ea1f76b437daab76",
                "responses": {
                    "200": {
                        "description": "Liste des reparations recuperee avec succes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reparations"
                ],
                "summary": "Creer une reparation",
                "operationId": "5c96d42f3a8b68ae9274730b2f3494b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicule_id",
                                    "mecanicien_id",
                                    "date_reparation",
                                    "description",
                                    "cout",
                                    "statut"
                                ],
                                "properties": {
                                    "vehicule_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mecanicien_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date_reparation": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-12"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Vidange moteur et controle des freins"
                                    },
                                    "cout": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250000
                                    },
                                    "statut": {
                                        "type": "string",
                                        "example": "planifiee",
                                        "enum": [
                                            "planifiee",
                                            "en_cours",
                                            "terminee",
                                            "annulee"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reparation creee avec succes"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/reparations/{id}": {
            "get": {
                "tags": [
                    "Reparations"
                ],
                "summary": "Afficher une reparation",
                "operationId": "a02d0f0bf99ad91fea66fcc2df4d2dac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de la reparation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reparation recuperee avec succes"
                    },
                    "404": {
                        "description": "Reparation introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reparations"
                ],
                "summary": "Modifier une reparation",
                "operationId": "606b0f2367dd4d89afcae2b9532a7d6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de la reparation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicule_id",
                                    "mecanicien_id",
                                    "date_reparation",
                                    "description",
                                    "cout",
                                    "statut"
                                ],
                                "properties": {
                                    "vehicule_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mecanicien_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date_reparation": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-13"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Remplacement des plaquettes de frein"
                                    },
                                    "cout": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 300000
                                    },
                                    "statut": {
                                        "type": "string",
                                        "example": "en_cours",
                                        "enum": [
                                            "planifiee",
                                            "en_cours",
                                            "terminee",
                                            "annulee"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reparation modifiee avec succes"
                    },
                    "404": {
                        "description": "Reparation introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reparations"
                ],
                "summary": "Supprimer une reparation",
                "operationId": "e0ae3bc8dfd8e1442bc73475d62e1cbe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de la reparation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reparation supprimee avec succes"
                    },
                    "404": {
                        "description": "Reparation introuvable"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Reparations"
                ],
                "summary": "Modifier partiellement une reparation",
                "operationId": "0f0fd715aa39c651bfaa4f77f620a47a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de la reparation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "vehicule_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mecanicien_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date_reparation": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-13"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Remplacement des plaquettes de frein"
                                    },
                                    "cout": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 300000
                                    },
                                    "statut": {
                                        "type": "string",
                                        "example": "terminee",
                                        "enum": [
                                            "planifiee",
                                            "en_cours",
                                            "terminee",
                                            "annulee"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reparation modifiee avec succes"
                    },
                    "404": {
                        "description": "Reparation introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
    },
...
}